chore: introduce "dynatrace" package
moved Client "factory" and other dynatrace related functions from cmdutils to dynatrace package
