+++ cmd/monaco/delete/delete.go
@@ -17,7 +17,7 @@ package delete
 import (
 	"errors"
 	"fmt"
-	"github.com/dynatrace/dynatrace-configuration-as-code/cmd/monaco/cmdutils"
+	"github.com/dynatrace/dynatrace-configuration-as-code/cmd/monaco/dynatrace"
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/errutils"
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/log"
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/maps"
@@ -82,7 +82,7 @@ func deleteConfigs(environments []manifest.EnvironmentDefinition, apis api.APIs,
 }
 
 func deleteConfigForEnvironment(env manifest.EnvironmentDefinition, apis api.APIs, entriesToDelete map[string][]delete.DeletePointer) []error {
-	dynatraceClient, err := cmdutils.CreateDTClient(env.URL.Value, env.Auth, false)
+	dynatraceClient, err := dynatrace.CreateClient(env.URL.Value, env.Auth, false)
 
 	if err != nil {
 		return []error{
