+++ cmd/monaco/dynatrace/dynatrace_test.go
@@ -1,5 +1,3 @@
-//go:build unit
-
 /*
  * @license
  * Copyright 2023 Dynatrace LLC
@@ -16,7 +14,7 @@
  * limitations under the License.
  */
 
-package cmdutils
+package dynatrace
 
 import (
 	"encoding/json"
