+++ cmd/monaco/integrationtest/integration_test_utils.go
@@ -19,7 +19,7 @@
 package integrationtest
 
 import (
-	"github.com/dynatrace/dynatrace-configuration-as-code/cmd/monaco/cmdutils"
+	"github.com/dynatrace/dynatrace-configuration-as-code/cmd/monaco/dynatrace"
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/testutils"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/api"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/client"
@@ -34,7 +34,7 @@ import (
 
 func CreateDynatraceClient(t *testing.T, environment manifest.EnvironmentDefinition) client.Client {
 
-	c, err := cmdutils.CreateDTClient(environment.URL.Value, environment.Auth, false)
+	c, err := dynatrace.CreateClient(environment.URL.Value, environment.Auth, false)
 	assert.NilError(t, err, "failed to create test client")
 
 	return c
